Adds two new configuation options:   secondary_host   secondary_port
If these are set then if calling a queue fails will try to post to a queue on a different host/port.

If the queue is reliable then it will keep trying the different queues (switching between primary and seconary hosts) until it connects correctly.
If the queue is unreliable, then it will try to connect to the secondary queue if the primary fails, and then will just completely fail.

Signed-off-by: Sean O'Halpin <sean.ohalpin@gmail.com>